Now all CATs can use GameplayScene.ts without cyclical dependency issue. Phew!
