+++ src/MODScript/EventHandler.ts
@@ -2,12 +2,12 @@ import { UpdateHandler } from "engine/MessageHandlers";
 import { EventDefinition } from "./MODscriptDefs";
 import { MODscriptEvent } from "./MODscriptEvent";
 import { BodyHandle } from "engine/BodyHandle";
+import { GameplayScene } from "engine/GameplayScene";
 
 export class EventHandler implements UpdateHandler {
 
     dummyData: string = "";
     events: MODscriptEvent[] = [];
-    playerReference: BodyHandle | undefined;
 
     private static _instance: EventHandler
     public static get instance(): EventHandler {
@@ -16,9 +16,8 @@ export class EventHandler implements UpdateHandler {
         return EventHandler._instance;
     }
 
-    initialize(playerReference: BodyHandle | undefined): void {
+    initialize(): void {
         //Filling in Dummy Data 
-        this.playerReference = playerReference;   
         this.events = this.createDummyData();
         for(let event of this.events)
             event.setCATs();
